services/vanna/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
asyncpg==0.29.0
pydantic==2.5.0
vanna==0.3.4
pandas==2.1.4
psycopg2-binary==2.9.9
